+++ app/frontend.py
@@ -6,6 +6,7 @@ from urllib.parse import urlparse
 
 from flask import (
     Flask,
+    Response,
     abort,
     make_response,
     redirect,
@@ -44,11 +45,10 @@ def create_app():
 
     @app.route("/")
     def index():
-        number_of_columns = request.cookies.get("number_of_columns", "2")
-        number_of_columns = int(number_of_columns) if number_of_columns.isdigit() else 0
-        refresh_period = request.cookies.get("refresh_period", "30")
-        refresh_period = int(refresh_period) if refresh_period.isdigit() else 0
-        cameras = wb.get_cameras(urlparse(request.root_url).hostname)
+        columns = request.cookies.get("number_of_columns", "2")
+        number_of_columns = int(columns) if columns.isdigit() else 0
+        refresh = request.cookies.get("refresh_period", "30")
+        refresh_period = int(refresh) if refresh.isdigit() else 0
         show_video = bool(request.cookies.get("show_video"))
         if "video" in request.args:
             show_video = True
@@ -57,14 +57,12 @@ def create_app():
         resp = make_response(
             render_template(
                 "index.html",
-                cameras=cameras,
-                enabled=len([cam for cam in cameras.values() if cam.get("enabled")]),
+                cameras=wb.get_cameras(urlparse(request.root_url).hostname),
                 number_of_columns=number_of_columns,
                 refresh_period=refresh_period,
                 hass=wb.hass,
                 version=wb.version,
                 show_video=show_video,
-                on_demand=wb.on_demand,
             )
         )
         resp.set_cookie("number_of_columns", str(number_of_columns))
@@ -74,38 +72,39 @@ def create_app():
 
     @app.route("/events/<path:event>/<path:cam>")
     def events(event: str, cam: str):
-        if event == "start":
-            try:
-                wb.start_on_demand(cam)
-            except StopIteration:
-                abort(404)
+        if event == "start" and not wb.start_on_demand(cam):
+            abort(404)
         return {}
 
     @app.route("/cameras")
-    def cameras():
-        return wb.get_cameras(
-            urlparse(request.root_url).hostname
-        )  # return json, for introspection or for future ajax UI
+    @app.route("/cameras/<path:cam_name>")
+    @app.route("/cameras/<path:cam_name>/<path:cam_cmd>")
+    def cameras(cam_name=None, cam_cmd=None):
+        """JSON api endpoints."""
+        if cam_name == "sse_status":  # Server Side Event
+            return Response(wb.sse_status(), mimetype="text/event-stream")
+
+        host = urlparse(request.root_url).hostname
+        if cam_name and cam_cmd == "status":
+            return {"status": wb.get_cam_status(cam_name)}
+        if cam_name:
+            return wb.get_cam_info(cam_name, host)
+        return wb.get_cameras(host)
 
     @app.route("/snapshot/<path:img_file>")
     def rtsp_snapshot(img_file: str):
         """Use ffmpeg to take a snapshot from the rtsp stream."""
         uri = Path(img_file).stem
-        if uri not in wb.get_cameras():
+        if not wb.rtsp_snap(uri, wait=True):
             abort(404)
-        wb.rtsp_snap(uri, wait=True)
         return send_from_directory(wb.img_path, img_file)
 
     @app.route("/photo/<path:img_file>")
     def boa_photo(img_file: str):
         """Take a photo on the camera and grab it over the boa http server."""
         uri = Path(img_file).stem
-        if not (cam := wb.get_cameras().get(uri)):
-            abort(404)
-        if photo := wb.boa_photo(cam.get("nickname")):
-            file = cam["nickname"] + "_" + photo[0]
-            return send_from_directory(wb.img_path, file)
-
+        if photo := wb.boa_photo(uri):
+            return send_from_directory(wb.img_path, uri + "_" + photo[0])
         return redirect(f"/img/{img_file}", code=302)
 
     @app.route("/img/<path:img_file>")
@@ -127,14 +126,14 @@ def create_app():
         """
         if restart_cmd == "cameras":
             wb.stop_cameras()
-            wb.start()
+            wb.run()
         elif restart_cmd == "rtsp_server":
             wb.stop_rtsp_server()
             wb.start_rtsp_server()
         elif restart_cmd == "all":
             wb.stop_cameras()
             wb.stop_rtsp_server()
-            wb.start()
+            wb.run()
             restart_cmd = "cameras,rtsp_server"
         else:
             return {"result": "error"}